add all types in utils directory in a centrelized way
