--- web/src/types/user.types.ts
@@ -1,11 +1,0 @@
-export interface User {
-    id: string;
-    email: string;
-    fullName: string;
-    phone?: string;
-    dateOfBirth?: Date;
-    isActive: boolean;
-    role: 'GUEST' | 'HOST' | 'ADMIN';
-    createdAt: string;
-    updatedAt: string;
-}
